--- libs/common/src/utils/dtools/validation/schemas/binarySchema.ts
@@ -1,5 +1,0 @@
-import { z } from 'zod'
-
-export const binarySchema = z.string().regex(/^[01]*$/, {
-  message: 'The value must be a valid, binary number'
-})
