--- libs/common/src/utils/dtools/validation/schemas/stringSchema.ts
@@ -1,3 +1,0 @@
-import { z } from 'zod'
-
-export const stringSchema = z.string()
