--- libs/common/src/utils/dtools/validation/validators/numberValidator.ts
@@ -1,9 +1,0 @@
-import { zodResultMessage } from '../../../../src/misc/zodResultMessage'
-import { decimalSchema } from '../schemas/decimalSchema'
-import { ValidatorResult } from './result'
-
-export function numberValidator (newValue: string): ValidatorResult {
-  const validated = decimalSchema.safeParse(newValue)
-  if (validated.success) return { success: true }
-  else return { success: false, error: zodResultMessage(validated) }
-}
