--- libs/common/src/utils/dtools/validation/validators/result.d.ts
@@ -1,3 +1,0 @@
-export type ValidatorResult =
-  | { success: true }
-  | { success: false; error: string };
